pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "garden-tasks-api"
version = "0.1.0"
description = "Backend API for community garden task coordination"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110,<1.0",
    "google-api-python-client>=2.194,<3.0",
    "google-auth>=2.49,<3.0",
    "uvicorn>=0.29,<1.0",
    "python-dotenv>=1.0,<2.0",
    "requests>=2.33,<3.0",
]

[dependency-groups]
dev = [
    "pytest>=9,<10",
    "ruff>=0.15,<1.0",
    "mypy>=1.19,<2.0",
    "types-requests>=2.33,<3.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = "src"
explicit_package_bases = true
